v2: finished schematics, start routing
The schematics for v2 should be mostly finished by now.
The new on/off button is ready and tested with a breadboard setup.
It now uses two FETs for a self-holding circuit for turning the
device on and off. To reach a safe logic level for Pico 2, a
2.7V Zener diode is used together with a pull-down for E9 affected
RP2350 revisions.
Besides, most pull-up resistors are 10k now to keep the BoM
smaller.

Routing started with a major redesign.