Add an voltage bypass mode when USB-C is connected
When USB-C is connected, the device should be charging, but also powered,
without taking care of the on/off button's state.
Otherwise, flashing the device via USB-C would only be
possible as long as the button is pressed which makes the usage
experience a mess and flaky.
This uses p-MOSFETs to do the switch and
a) avoid a battery voltage going to the system when USB-C is connected
b) avoid powering having both, USB-C and Micro-USB active

TODO: This adds the basic idea, routing must be fixed!